Test that an HTTP request is made with a header

This expectation checks that a HTTP header (and potentially header value) is present in a request. It works by inspecting the request object and raising warnings that are caught by testthat::expect_warning().

expect_header(..., ignore.case = TRUE)

Arguments

...
Arguments passed to expect_warning

ignore.case
logical: if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching. Default is TRUE; note that this is the opposite of expect_warning but is appropriate here because HTTP header names are case insensitive.

Value

NULL, according to expect_warning.

Details

expect_header works both in the mock HTTP contexts and on "live" HTTP requests.

Examples

library(httr)
with_fake_http({
    expect_header(GET("http://example.com", config=add_headers(Accept="image/png")),
        "Accept: image/png")
})

#> Loading required package: curl

#> 
#> Attaching package: ‘curl’

#> The following object is masked from ‘package:httr’:
#> 
#>     handle_reset

#> Tracing function "form_file" as seen from package "httr"

#> GET http://example.com

#> Untracing function "form_file" in package "curl"

#> Response [http://example.com]
#>   Date: 2020-02-03 12:22
#>   Status: 200
#>   Content-Type: <unknown>
#>   Size: 4 B
#> <BINARY BODY>