Return something that looks enough like an httr 'response'

These functions allow mocking of HTTP requests without requiring an internet connection or server to run against. Their return shape is a 'httr' "response" class object that should behave like a real response generated by a real request.

fakeResponse(request, verb = "GET", status_code = 200, headers = list(),
  content = NULL)

Arguments

request
An 'httr' request-class object. A character URL is also accepted, for which a fake request object will be created, using the verb argument as well. This behavior is deprecated.

verb
Character name for the HTTP verb. Default is "GET". This argument is deprecated; instead you should pass in a request

status_code
Integer HTTP response status

headers
Optional list of additional response headers to return

content
If supplied, a JSON-serializable list that will be returned as response content with Content-Type: application/json. If no content is provided, and if the status_code is not 204 No Content, the url will be set as the response content with Content-Type: text/plain.

Value

An 'httr' response class object.