Set an alternate directory for mock API fixtures

By default, with_mock_api will look for mocks relative to the current working directory (the test directory). If you want to look in other places, you can call .mockPaths to add directories to the search path.

.mockPaths(new)

Arguments

new
Either a character vector of path(s) to add, or NULL to reset to the default.

Value

If new is omitted, the function returns the current search paths, a a character vector. If new is provided, the updated value will be returned invisibly.

Details

It works like base::.libPaths(): any directories you specify will be added to the list and searched first. The default directory will be searched last. Only unique values are kept: if you provide a path that is already found in .mockPaths, the result effectively moves that path to the first position.

For finer-grained control, or to completely override the default behavior of searching in the current working directory, you can set the option "httptest.mock.paths" directly.

Examples

identical(.mockPaths(), ".")

#> [1] TRUE

.mockPaths("/var/somewhere/else")
identical(.mockPaths(), c("/var/somewhere/else", "."))

#> [1] TRUE

.mockPaths(NULL)
identical(.mockPaths(), ".")

#> [1] TRUE