Remove sensitive content from HTTP responses

When recording requests for use as test fixtures, you don't want to include secrets like authentication tokens and personal ids. These functions provide a means for redacting this kind of content, or anything you want, from responses that capture_requests() saves.

redact_auth(response)

redact_cookies(response)

redact_headers(headers = c())

redact_HTTP_auth(response)

within_body_text(FUN)

Arguments

response
An 'httr' response object to sanitize. Redacting functions should take as their only argument a response object.

headers
For redact_headers(), a character vector of header names to sanitize. Note that redact_headers() itself does not do redacting but returns a function that when called does the redacting.

FUN
For within_body_text(), a function that takes as its argument a character vector and returns a modified version of that. This function will be applied to the text of the response's "content".

Value

Redacting functions must return a well-formed 'httr' response object. redact_auth(), redact_HTTP_auth(), and redact_cookies() themselves are redacting functions, while redact_headers() and within_body_text() return redacting functions.

Details

redact_cookies() removes cookies from 'httr' response objects. redact_headers() lets you target selected request and response headers for redaction. redact_HTTP_auth() removes username:password-based HTTP auth credentials. redact_auth() is a convenience wrapper around them for a useful default redactor in capture_requests(). within_body_text()

See also

vignette("redacting", package="httptest") for a detailed discussion of what these functions do and how to customize them.