Write out a captured response

Write out a captured response

Usage

save_response(response, simplify = TRUE)

Arguments

response

An 'httr' response object

simplify

logical: if TRUE (default), JSON responses with status 200 and a supported Content-Type will be written as just the text of the response body. In all other cases, and when simplify is FALSE, the "response" object will be written out to a .R file using base::dput().

Value

The character file name that was written out